+++ lib/nosedrum/invoker/split.ex
@@ -23,9 +23,13 @@ defmodule Nosedrum.Invoker.Split do
   alias Nostrum.Struct.Message
 
   @doc false
-  def handle_message(message, storage \\ Nosedrum.Storage.ETS) do
+  def handle_message(
+        message,
+        storage \\ Nosedrum.Storage.ETS,
+        storage_process \\ :nosedrum_commands
+      ) do
     with [@prefix <> command | args] <- Helpers.quoted_split(message.content),
-         cog when cog != nil <- storage.lookup_command(command) do
+         cog when cog != nil <- storage.lookup_command(command, storage_process) do
       handle_command(cog, message, args)
     else
       _mismatch -> :ignored